chore: restructure socket logic
